add newsletter email validation

+++ inc/Hasnain/REST/Newsletter.php
@@ -13,11 +13,20 @@
     public static function create()
     {
       $option = self::$option;
+      $email = $_POST['email'];
       $subscriptions = get_option($option, []);
-      $updated_subscriptions = [...$subscriptions, $_POST['email']];
 
+      if (!filter_var($email, FILTER_VALIDATE_EMAIL)) {
+        return wp_send_json(['message' => 'Are you sure this is a valid email address?'], 400);
+      }
+
+      if (in_array($email, $subscriptions)) {
+        return wp_send_json(['message' => 'This email address has already subscribed!'], 400);
+      }
+
+      $updated_subscriptions = [...$subscriptions, $_POST['email']];
       update_option($option, $updated_subscriptions);
 
-      wp_send_json(['message' => 'created'], 200);
+      return wp_send_json(['message' => 'Thank you for subscribing to our newsletter!'], 200);
     }
   }
